requirements.txt
fastapi[all]
fastapi-sqlalchemy 
pydantic 
alembic 
uvicorn 
python-dotenv
psycopg2-binary